KIECLOUD-81] - Add support to validate ImageStreams with template-validator-tool
