+++ tools/openshift-template-validator/glide.lock
@@ -1,5 +1,5 @@
-hash: a0461b2a0672e6ba3849e92dd8b6430717d03f40af2305ee813dd4856d12a9d2
-updated: 2018-09-20T17:33:19.71713816-03:00
+hash: 6f4fc795a47e933aa6f1753ffd277763fcfb85ed629eabdf046cb11ba083a923
+updated: 2018-12-27T19:00:31.80750195-02:00
 imports:
 - name: github.com/asaskevich/govalidator
   version: ccb8e960c48f04d6935e72476ae4a51028f9e22f
@@ -17,15 +17,15 @@ imports:
   - digestset
   - reference
 - name: github.com/docker/docker
-  version: 5d848e66df951faffad93d70d8d31470af4cc5b0
+  version: 4f3616fb1c112e206b88cb7a9922bf49067a7756
   subpackages:
   - api/types
 - name: github.com/docker/go-connections
-  version: 97c2040d34dfae1d1b1275fa3a78dbdd2f41cf7e
+  version: 3ede32e2033de7505e6500d6c868c2b9ed9f169d
 - name: github.com/docker/go-units
-  version: 47565b4f722fb6ceae66b95f853feed578a4a51c
+  version: 9e638d38cf6977a37a8ea0078f3ee75a7cdb2dd1
 - name: github.com/getsentry/raven-go
-  version: 084a9de9eb0361fbd5ded14b55c84e5493a5d7f6
+  version: 32a13797442ccb601b11761d74232773c1402d14
 - name: github.com/ghodss/yaml
   version: 73d445a93680fa1a78ae23a5839bad48f32ba1ee
 - name: github.com/gogo/protobuf
@@ -36,23 +36,23 @@ imports:
 - name: github.com/golang/glog
   version: 44145f04b68cf362d9c4df2182967c2275eaefed
 - name: github.com/golang/protobuf
-  version: 7716a980bceefe9a0213654823d9ea62d23b87a2
+  version: b4deda0973fb4c70b50d226b1af49f3da59f5265
 - name: github.com/google/gofuzz
   version: 44d81051d367757e1c7c6a5a86423ece9afcf63c
 - name: github.com/googleapis/gnostic
-  version: 48a0ecefe2e4190c7a2d63b477875854a2f993b3
+  version: 0c5108395e2debce0d731cf0287ddf7242066aba
   subpackages:
   - OpenAPIv2
 - name: github.com/hashicorp/golang-lru
-  version: 20f1fb78b0740ba8c3cb143a61e86ba5c8669768
+  version: a0d98a5f288019575c6d1f4bb1573fef2d1fcdc4
 - name: github.com/imdario/mergo
-  version: 9f23e2d6bd2a77f959b2bf6acdbefd708a83a4a4
+  version: 6633656539c1639d9d78127b7d47c622b5d7b6dc
 - name: github.com/inconshreveable/mousetrap
   version: 76626ae9c91c4f2a10f34cad8ce83ea42c93bb75
 - name: github.com/json-iterator/go
   version: f2b4162afba35581b6d4a50d3b8f34e33c144682
 - name: github.com/konsorten/go-windows-terminal-sequences
-  version: b729f2633dfe35f4d1d8a32385f6685610ce1cb5
+  version: 5c8c8bd35d3832f5d134ae1e1e375b69a4d25242
 - name: github.com/modern-go/concurrent
   version: bacd9c7ef1dd9b15be4a9909b8ac7a4e313eec94
 - name: github.com/modern-go/reflect2
@@ -60,11 +60,11 @@ imports:
 - name: github.com/opencontainers/go-digest
   version: a6d0ee40d4207ea02364bd3b9e8e77b9159ba1eb
 - name: github.com/opencontainers/image-spec
-  version: 7b1e489870acb042978a3935d2fb76f8a79aff81
+  version: 372ad780f63454fbbbbcc7cf80e5b90245c13e13
   subpackages:
   - specs-go/v1
 - name: github.com/openshift/api
-  version: 4507dcb6a81f6593aa336c5b47927cf0ae297b72
+  version: 4703f3e71d833812a81e0fa3a2da1257e3efd85c
   subpackages:
   - apps/v1
   - authorization/v1
@@ -81,19 +81,23 @@ imports:
 - name: github.com/openshift/library-go
   version: 0b8367a4679859036c27a30dbe010d76409e7075
 - name: github.com/pkg/errors
-  version: c059e472caf75dbe73903f6521a20abac245b17f
+  version: 645ef00459ed84a119197bfb8d8205042c6df63d
 - name: github.com/pkg/profile
   version: 057bc52a47ec3c79498dda63f4a6f8298725e976
 - name: github.com/sirupsen/logrus
-  version: 33a1e118e113c7d1dd24a80f80670864cd598519
+  version: 89742aefa4b206dcf400792f3bd35b542998eb3b
 - name: github.com/spf13/cobra
-  version: 8d114be902bc9f08717804830a55c48378108a28
+  version: c439c4fa093711d42e1b01acb1235b52004753c1
 - name: github.com/spf13/pflag
   version: 583c0c0531f06d5278b7d917446061adc344b5cd
+- name: github.com/stretchr/testify
+  version: 660f15d67dbb878de0d9d79894f728d691c91b91
+  subpackages:
+  - assert
 - name: github.com/urfave/cli
   version: cfb38830724cc34fedffe9a2a29fb54fa9169cd1
 - name: golang.org/x/crypto
-  version: 0e37d006457bf46f9e6692014ba72ef82c33022c
+  version: de0752318171da717af4ce24d0a2e8626afaeb11
   subpackages:
   - ssh/terminal
 - name: golang.org/x/net
@@ -104,7 +108,7 @@ imports:
   - idna
   - lex/httplex
 - name: golang.org/x/sys
-  version: d641721ec2dead6fe5ca284096fe4b1fcd49e427
+  version: 95c6576299259db960f6c5b9b69ea52422860fce
   subpackages:
   - unix
 - name: golang.org/x/text
@@ -115,16 +119,17 @@ imports:
   - unicode/bidi
   - unicode/norm
 - name: golang.org/x/time
-  version: fbb02b2291d28baffd63558aa44b4b56f178d650
+  version: f51c12702a4d776e4c1fa9b0fabab841babae631
   subpackages:
   - rate
 - name: gopkg.in/inf.v0
   version: 3887ee99ecf07df5b447e9b00d9c0b2adaa9f3e4
 - name: gopkg.in/yaml.v2
-  version: 5420a8b6744d3b0345ab293f6fcba19c978f1183
+  version: 51d6538a90f86fe93ac480b35f37b2be17fef232
 - name: k8s.io/api
   version: 2d6f90ab1293a1fb871cf149423ebb72aa7423aa
   subpackages:
+  - apps/v1
   - apps/v1beta1
   - core/v1
   - extensions/v1beta1
@@ -140,6 +145,7 @@ imports:
   - pkg/api/meta
   - pkg/api/resource
   - pkg/api/validation
+  - pkg/api/validation/path
   - pkg/apis/meta/internalversion
   - pkg/apis/meta/v1
   - pkg/apis/meta/v1/unstructured
@@ -182,11 +188,11 @@ imports:
 - name: k8s.io/kube-aggregator
   version: 89cd614e9090a2f1e78316ed459857c49c55d276
 - name: k8s.io/kube-openapi
-  version: e3762e86a74c878ffed47484592986685639c2cd
+  version: 8a9b82f00b3a86eac24681da3f9fe6c34c01cea2
   subpackages:
   - pkg/util/proto
 - name: k8s.io/kubernetes
-  version: f818aa3443bc9820a54afb479a553e5cc3c4dc3a
+  version: 21afb3cbc1a9669fc6fb7c7ccb70192fa0e10934
   subpackages:
   - pkg/api/legacyscheme
   - pkg/api/service
@@ -200,6 +206,8 @@ imports:
   - pkg/apis/extensions
   - pkg/apis/networking
   - pkg/apis/policy
+  - pkg/apis/rbac
+  - pkg/apis/rbac/validation
   - pkg/apis/scheduling
   - pkg/capabilities
   - pkg/features
@@ -211,4 +219,8 @@ imports:
   - pkg/util/net/sets
   - pkg/util/parsers
   - pkg/util/pointer
-testImports: []
+testImports:
+- name: github.com/pmezard/go-difflib
+  version: d8ed2627bdf02c080bf22230dbb337003b7aba2d
+  subpackages:
+  - difflib
